Allow only Uint8Array for screenshots
